Added methods for item use and status effects
Code for changing attack speed, adding burn, and changing enemy movement speed still needs to be made, but setting the effects on the player all works, as well as the timer for removing it.

Boolean array also is here for HUD use, to just have to display the correct status icons whenever the status is true.